fix: Shadowing 점수 반환 수정(0.5점 단위로 출력되도록)

+++ BackEnd/SpringBootServer/src/main/java/com/ssafy/backend/domain/news/dto/ShadowingResponse.java
@@ -1,4 +1,4 @@
 package com.ssafy.backend.domain.news.dto;
 
-public record ShadowingResponse(float similarity) {
+public record ShadowingResponse(double similarity) {
 }
